requirements.txt
requests==2.26.0
pandas==1.3.3
sqlalchemy==1.4.23
azure-keyvault-secrets==4.3.0
azure-identity==1.6.4
azure-storage-blob==12.9.0